final-pjt-back/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.7.0
pytz==2022.6
requests==2.30.0
sqlparse==0.4.3
urllib3==2.0.2
